+++ lib/core/router.dart
@@ -5,14 +5,14 @@ import 'package:mongbi_app/core/main_scaffold.dart';
 import 'package:mongbi_app/core/responsive_layout.dart';
 import 'package:mongbi_app/core/route_observer.dart';
 import 'package:mongbi_app/features/auth/presentation/pages/social_login_page.dart';
+import 'package:mongbi_app/features/dream/presentation/pages/dream_analysis/dream_analysis_loading_page.dart';
+import 'package:mongbi_app/features/dream/presentation/pages/dream_analysis/dream_analysis_result_page.dart';
+import 'package:mongbi_app/features/dream/presentation/pages/dream_analysis/dream_interpretation_page.dart';
+import 'package:mongbi_app/features/dream/presentation/pages/dream_write/dream_intro_page.dart';
+import 'package:mongbi_app/features/dream/presentation/pages/dream_write/dream_write_page.dart';
 import 'package:mongbi_app/presentation/alarm/alarm_page.dart';
 import 'package:mongbi_app/presentation/challenge/challenge_intro_page.dart';
 import 'package:mongbi_app/presentation/challenge/challenge_page.dart';
-import 'package:mongbi_app/presentation/dream/dream_analysis_loading_page.dart';
-import 'package:mongbi_app/presentation/dream/dream_analysis_result_page.dart';
-import 'package:mongbi_app/presentation/dream/dream_interpretation_page.dart';
-import 'package:mongbi_app/presentation/dream/dream_intro_page.dart';
-import 'package:mongbi_app/presentation/dream/dream_write_page.dart';
 import 'package:mongbi_app/presentation/history/history_page.dart';
 import 'package:mongbi_app/presentation/home/home_page.dart';
 import 'package:mongbi_app/presentation/onboarding/onbording_exit_page.dart';
