--- lib/data/data_sources/dream_analysis_data_source.dart
@@ -1,3 +1,0 @@
-abstract interface class DreamAnalysisDataSource {
-  Future<Map<String, dynamic>> analyzeDream(String dreamContent, int dreamScore);
-}
