+++ lib/features/dream/data/data_sources/dream_analysis_data_source.dart
@@ -1,0 +1,6 @@
+abstract interface class DreamAnalysisDataSource {
+  Future<Map<String, dynamic>> analyzeDream(
+    String dreamContent,
+    int dreamScore,
+  );
+}
