+++ lib/features/dream/data/data_sources/dream_save_data_source.dart
@@ -1,4 +1,4 @@
-import 'package:mongbi_app/data/dtos/dream_dto.dart';
+import 'package:mongbi_app/features/dream/data/dtos/dream_dto.dart';
 
 abstract interface class DreamSaveDataSource {
   Future<int> saveDream(DreamDto dream);
