+++ lib/features/dream/data/data_sources/remote_dream_data_source.dart
@@ -1,6 +1,6 @@
 import 'package:dio/dio.dart';
-import 'package:mongbi_app/data/data_sources/dream_save_data_source.dart';
-import 'package:mongbi_app/data/dtos/dream_dto.dart';
+import 'package:mongbi_app/features/dream/data/data_sources/dream_save_data_source.dart';
+import 'package:mongbi_app/features/dream/data/dtos/dream_dto.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 
 class RemoteDreamDataSource implements DreamSaveDataSource {
@@ -16,12 +16,14 @@ class RemoteDreamDataSource implements DreamSaveDataSource {
       if (response.statusCode == 201 && response.data['success'] == true) {
         return response.data['data']['DREAM_IDX'] as int;
       } else {
-        final error = Exception(response.data['message'] ?? '알 수 없는 오류가 발생하였습니다.');
-        await Sentry.captureException(error); 
+        final error = Exception(
+          response.data['message'] ?? '알 수 없는 오류가 발생하였습니다.',
+        );
+        await Sentry.captureException(error);
         throw error;
       }
     } on DioException catch (e, s) {
-      await Sentry.captureException(e, stackTrace: s); 
+      await Sentry.captureException(e, stackTrace: s);
       throw Exception(e.message ?? '네트워크 오류가 발생하였습니다.');
     } catch (e, s) {
       await Sentry.captureException(e, stackTrace: s);
