+++ lib/features/dream/data/repositories/remote_dream_repository.dart
@@ -1,9 +1,9 @@
-import 'package:mongbi_app/data/data_sources/dream_analysis_data_source.dart';
-import 'package:mongbi_app/data/data_sources/dream_check_data_source.dart';
-import 'package:mongbi_app/data/data_sources/dream_save_data_source.dart';
-import 'package:mongbi_app/data/dtos/dream_dto.dart';
-import 'package:mongbi_app/domain/entities/dream.dart';
-import 'package:mongbi_app/domain/repositories/dream_repository.dart';
+import 'package:mongbi_app/features/dream/data/data_sources/dream_analysis_data_source.dart';
+import 'package:mongbi_app/features/dream/data/data_sources/dream_check_data_source.dart';
+import 'package:mongbi_app/features/dream/data/data_sources/dream_save_data_source.dart';
+import 'package:mongbi_app/features/dream/data/dtos/dream_dto.dart';
+import 'package:mongbi_app/features/dream/domain/entity/dream.dart';
+import 'package:mongbi_app/features/dream/domain/repositoreis/dream_repository.dart';
 
 class RemoteDreamRepository implements DreamRepository {
   RemoteDreamRepository(
