+++ lib/features/dream/domain/use_cases/analyze_and_save_dream_use_case.dart
@@ -1,6 +1,6 @@
-import 'package:mongbi_app/domain/entities/dream.dart';
-import 'package:mongbi_app/domain/use_cases/analyze_dream_use_case.dart';
-import 'package:mongbi_app/domain/use_cases/save_dream_use_case.dart';
+import 'package:mongbi_app/features/dream/domain/entity/dream.dart';
+import 'package:mongbi_app/features/dream/domain/use_cases/analyze_dream_use_case.dart';
+import 'package:mongbi_app/features/dream/domain/use_cases/save_dream_use_case.dart';
 
 class AnalyzeAndSaveDreamUseCase {
   AnalyzeAndSaveDreamUseCase(this.analyzeDreamUseCase, this.saveDreamUseCase);
