+++ lib/features/dream/domain/use_cases/can_write_dream_today_use_case.dart
@@ -1,4 +1,4 @@
-import 'package:mongbi_app/domain/repositories/dream_repository.dart';
+import 'package:mongbi_app/features/dream/domain/repositoreis/dream_repository.dart';
 
 class CanWriteDreamTodayUseCase {
   CanWriteDreamTodayUseCase({required this.repository});
