+++ lib/features/dream/presentation/pages/dream_analysis/dream_interpretation_page.dart
@@ -3,10 +3,10 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:go_router/go_router.dart';
 import 'package:mongbi_app/core/font.dart';
+import 'package:mongbi_app/features/dream/presentation/pages/dream_analysis/widgets/dream_section_card.dart';
+import 'package:mongbi_app/features/dream/presentation/pages/dream_analysis/widgets/mongbi_comment_card.dart';
+import 'package:mongbi_app/features/dream/presentation/widgets/custom_button.dart';
 import 'package:mongbi_app/presentation/common/action_button_row.dart';
-import 'package:mongbi_app/presentation/dream/widgets/custom_button.dart';
-import 'package:mongbi_app/presentation/dream/widgets/dream_section_card.dart';
-import 'package:mongbi_app/presentation/dream/widgets/mongbi_comment_card.dart';
 import 'package:mongbi_app/presentation/home/widgets/completion_bottom_sheet.dart';
 import 'package:mongbi_app/providers/dream_provider.dart';
