+++ lib/presentation/challenge/widgets/mongbi_dialog.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:mongbi_app/core/font.dart';
 import 'package:mongbi_app/core/responsive_layout.dart';
 import 'package:mongbi_app/features/auth/presentation/pages/widgets/mongbi_image_widget.dart';
-import 'package:mongbi_app/presentation/dream/widgets/custom_button.dart';
+import 'package:mongbi_app/features/dream/presentation/widgets/custom_button.dart';
 
 class MongbiDialog extends StatelessWidget {
   const MongbiDialog({
