+++ lib/providers/dream_provider.dart
@@ -1,20 +1,20 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:mongbi_app/data/data_sources/dream_analysis_data_source.dart';
-import 'package:mongbi_app/data/data_sources/dream_check_data_source.dart';
-import 'package:mongbi_app/data/data_sources/dream_save_data_source.dart';
-import 'package:mongbi_app/data/data_sources/remote_dream_analysis_data_source.dart';
-import 'package:mongbi_app/data/data_sources/remote_dream_check_data_source.dart';
-import 'package:mongbi_app/data/data_sources/remote_dream_data_source.dart';
-import 'package:mongbi_app/data/repositories/remote_dream_repository.dart';
-import 'package:mongbi_app/domain/repositories/dream_repository.dart';
-import 'package:mongbi_app/domain/use_cases/analyze_and_save_dream_use_case.dart';
-import 'package:mongbi_app/domain/use_cases/analyze_dream_use_case.dart';
-import 'package:mongbi_app/domain/use_cases/can_write_dream_today_use_case.dart';
-import 'package:mongbi_app/domain/use_cases/save_dream_use_case.dart';
-import 'package:mongbi_app/presentation/dream/models/dream_interpretation_state.dart';
-import 'package:mongbi_app/presentation/dream/models/dream_write_state.dart';
-import 'package:mongbi_app/presentation/dream/view_models/dream_interpretation_view_model.dart';
-import 'package:mongbi_app/presentation/dream/view_models/dream_write_view_model.dart';
+import 'package:mongbi_app/features/dream/data/data_sources/dream_analysis_data_source.dart';
+import 'package:mongbi_app/features/dream/data/data_sources/dream_check_data_source.dart';
+import 'package:mongbi_app/features/dream/data/data_sources/dream_save_data_source.dart';
+import 'package:mongbi_app/features/dream/data/data_sources/remote_dream_analysis_data_source.dart';
+import 'package:mongbi_app/features/dream/data/data_sources/remote_dream_check_data_source.dart';
+import 'package:mongbi_app/features/dream/data/data_sources/remote_dream_data_source.dart';
+import 'package:mongbi_app/features/dream/data/repositories/remote_dream_repository.dart';
+import 'package:mongbi_app/features/dream/domain/repositoreis/dream_repository.dart';
+import 'package:mongbi_app/features/dream/domain/use_cases/analyze_and_save_dream_use_case.dart';
+import 'package:mongbi_app/features/dream/domain/use_cases/analyze_dream_use_case.dart';
+import 'package:mongbi_app/features/dream/domain/use_cases/can_write_dream_today_use_case.dart';
+import 'package:mongbi_app/features/dream/domain/use_cases/save_dream_use_case.dart';
+import 'package:mongbi_app/features/dream/presentation/models/dream_interpretation_state.dart';
+import 'package:mongbi_app/features/dream/presentation/models/dream_write_state.dart';
+import 'package:mongbi_app/features/dream/presentation/view_models/dream_interpretation_view_model.dart';
+import 'package:mongbi_app/features/dream/presentation/view_models/dream_write_view_model.dart';
 import 'package:mongbi_app/providers/core_providers.dart';
 
 final _dreamDataSourceProvider = Provider<DreamSaveDataSource>(
