+++ test/data/dream_repository_test.dart
@@ -1,11 +1,11 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
-import 'package:mongbi_app/data/data_sources/dream_analysis_data_source.dart';
-import 'package:mongbi_app/data/data_sources/dream_check_data_source.dart';
-import 'package:mongbi_app/data/data_sources/dream_save_data_source.dart';
-import 'package:mongbi_app/data/dtos/dream_dto.dart';
-import 'package:mongbi_app/data/repositories/remote_dream_repository.dart';
-import 'package:mongbi_app/domain/entities/dream.dart';
+import 'package:mongbi_app/features/dream/data/data_sources/dream_analysis_data_source.dart';
+import 'package:mongbi_app/features/dream/data/data_sources/dream_check_data_source.dart';
+import 'package:mongbi_app/features/dream/data/data_sources/dream_save_data_source.dart';
+import 'package:mongbi_app/features/dream/data/dtos/dream_dto.dart';
+import 'package:mongbi_app/features/dream/data/repositories/remote_dream_repository.dart';
+import 'package:mongbi_app/features/dream/domain/entity/dream.dart';
 
 class FakeDreamDto extends Fake implements DreamDto {}
